Fix crop for some case of portrait orientation
It seems that while Nitro considers a portrait photo to have H > W, in
Adobe's world they have W > H and an "orientation" attribute that says
the photo was rotated. As a result, we have to transform the crop
settings, e.g. "crs:CropTop" will actually be applied on the left edge
of a portrait photo, "crs:CropRight" will be on the top, etc.

There are probably other orientation modes to worry about, and I didn't
address angles at all, but this does show a path to correct it.