Update converter to find EditModel in exported XMP
Importing a Nitro RAW (CR3+XMP) into Lightroom and exporting it back out
is one step in the translator process. During this export process LR
changes the way the Nitro model is stored in the file. The change looks
to be compatible; I'm probably just doing something silly with my XML
parsing. But for now, make the edit to leave the original Nitro XMP
untouched, and just operate on the exported one.